refactor: tx commands and cli simplificated
