+++ moccasin/msig_cli/tx/helpers.py
@@ -1,0 +1,32 @@
+from prompt_toolkit import HTML, print_formatted_text
+from safe_eth.util.util import to_0x_hex_str
+from safe_eth.safe import SafeTx
+
+
+def pretty_print_safe_tx(safe_tx: SafeTx):
+    """Pretty-print SafeTx fields."""
+
+    print_formatted_text(HTML("<b><orange>SafeTx</orange></b>"))
+    print_formatted_text(HTML(f"\t<b><yellow>Nonce:</yellow></b> {safe_tx.safe_nonce}"))
+    print_formatted_text(HTML(f"\t<b><yellow>To:</yellow></b> {safe_tx.to}"))
+    print_formatted_text(HTML(f"\t<b><yellow>Value:</yellow></b> {safe_tx.value}"))
+    print_formatted_text(
+        HTML(f"\t<b><yellow>Data:</yellow></b> {to_0x_hex_str(safe_tx.data)}")
+    )
+    print_formatted_text(
+        HTML(f"\t<b><yellow>Operation:</yellow></b> {safe_tx.operation}")
+    )
+    print_formatted_text(
+        HTML(f"\t<b><yellow>SafeTxGas:</yellow></b> {safe_tx.safe_tx_gas}")
+    )
+    print_formatted_text(HTML(f"\t<b><yellow>BaseGas:</yellow></b> {safe_tx.base_gas}"))
+    print_formatted_text(
+        HTML(f"\t<b><yellow>GasPrice:</yellow></b> {safe_tx.gas_price}")
+    )
+    print_formatted_text(
+        HTML(f"\t<b><yellow>GasToken:</yellow></b> {safe_tx.gas_token}")
+    )
+    print_formatted_text(
+        HTML(f"\t<b><yellow>RefundReceiver:</yellow></b> {safe_tx.refund_receiver}")
+    )
+    print_formatted_text(HTML(f"\t<b><yellow>Signers:</yellow></b> {safe_tx.signers}"))
