+++ tests/cli/test_cli_msig.py
@@ -4,7 +4,7 @@ import os
 import pytest
 
 
-MOX_CMD = ["mox", "msig", "tx", "build"]
+MOX_CMD = ["mox", "msig", "tx_build"]
 
 
 def test_cli_tx_builder_interactive(temp_msig_workdir):
